Finish up KVM UEFI call test
The weird part is that when debugging is on, ARM64
seems to not debug exit on the instruction right
after the MMIO instruction.

Also it is seeming dicey to mess with the Pc anyway.
The best bet may be a 4 instruction sequence with,
basically, a no op after the MMIO, and no messing with
the PC? ouch.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>